Struct gists::data::Data

pub struct Data {
    pub creds: ArgonConfig,
    pub settings: Settings,
}

App data

Fields
creds: ArgonConfig

credential-procession policy

settings: Settings

settings

Implementations

impl Data

pub async fn email_exists(
&self,
db: &Box<dyn GistDatabase>,
email: &str
) -> ServiceResult<AccountCheckResp>

check if email exists on database

pub async fn set_email(
&self,
db: &Box<dyn GistDatabase>,
username: &str,
email: &str
) -> ServiceResult<()>

update email

pub async fn username_exists(
&self,
db: &Box<dyn GistDatabase>,
username: &str
) -> ServiceResult<AccountCheckResp>

check if email exists in database

pub async fn update_username(
&self,
db: &Box<dyn GistDatabase>,
current_username: &str,
new_username: &str
) -> ServiceResult<String>

update username of a registered user

pub async fn get_secret(
&self,
db: &Box<dyn GistDatabase>,
username: &str
) -> ServiceResult<Secret>

get account secret of a registered user

pub async fn update_user_secret(
&self,
db: &Box<dyn GistDatabase>,
username: &str
) -> ServiceResult<String>

update account secret of a registered user

async fn authenticate(
&self,
db: &Box<dyn GistDatabase>,
username: &str,
password: &str
) -> ServiceResult<()>

pub async fn delete_user(
&self,
db: &Box<dyn GistDatabase>,
username: &str,
password: &str
) -> ServiceResult<()>

delete user

pub async fn change_password(
&self,
db: &Box<dyn GistDatabase>,
username: &str,
payload: &ChangePasswordReqest
) -> ServiceResult<()>

change password

impl Data

pub async fn login<T: GistDatabase>(
&self,
db: &T,
payload: &Login
) -> ServiceResult<String>

Log in method. Returns Ok(()) when user is authenticated and errors when authentication fails

pub async fn register<T: GistDatabase>(
&self,
db: &T,
payload: &Register
) -> ServiceResult<()>

register new user

impl Data

pub async fn new_gist<T: GistDatabase>(
&self,
db: &T,
msg: &CreateGist<'_>
) -> ServiceResult<Gist>

pub(crate) fn get_repository_path(&self, gist_id: &str) -> PathBuf

pub async fn write_file<T: GistDatabase>(
&self,
_db: &T,
gist_id: &str,
files: &[File]
) -> ServiceResult<()>

pub async fn read_file<T: GistDatabase>(
&self,
_db: &T,
gist_id: &str,
path: &str
) -> ServiceResult<Vec<u8>>

Please note that this method expects path to not contain any spaces Use escape_spaces before calling this method

For example, a read request for “foo bar.md” will fail even if that file is present in the repository. However, it will succeed if the output of escape_spaces is used in the request.

impl Data

pub fn get_creds() -> ArgonConfig

Get credential-processing policy

pub fn new(settings: Option<Settings>) -> Arc<Self>

create new instance of app data